Update sound card handling based on MMDVM-UDRC changes from NH6Z.

+++ SoundCardReaderWriter.cpp
@@ -259,7 +259,7 @@ bool CSoundCardReaderWriter::open()
 		::fprintf(stderr, "Cannot set sample format (%s)\n", ::snd_strerror(err));
 		return false;
 	}
-	
+
 	if ((err = ::snd_pcm_hw_params_set_rate(playHandle, hw_params, m_sampleRate, 0)) < 0) {
 		::fprintf(stderr, "Cannot set sample rate (%s)\n", ::snd_strerror(err));
 		return false;
@@ -275,14 +275,14 @@ bool CSoundCardReaderWriter::open()
 			return false;
 		}
 	}
-	
+
 	if ((err = ::snd_pcm_hw_params(playHandle, hw_params)) < 0) {
 		::fprintf(stderr, "Cannot set parameters (%s)\n", ::snd_strerror(err));
 		return false;
 	}
-	
+
 	::snd_pcm_hw_params_free(hw_params);
-	
+
 	if ((err = ::snd_pcm_prepare(playHandle)) < 0) {
 		::fprintf(stderr, "Cannot prepare audio interface for use (%s)\n", ::snd_strerror(err));
 		return false;
@@ -304,7 +304,7 @@ bool CSoundCardReaderWriter::open()
 		::fprintf(stderr, "Cannot initialize hardware parameter structure (%s)\n", ::snd_strerror(err));
 		return false;
 	}
-	
+
 	if ((err = ::snd_pcm_hw_params_set_access(recHandle, hw_params, SND_PCM_ACCESS_RW_INTERLEAVED)) < 0) {
 		::fprintf(stderr, "Cannot set access type (%s)\n", ::snd_strerror(err));
 		return false;
@@ -314,14 +314,14 @@ bool CSoundCardReaderWriter::open()
 		::fprintf(stderr, "Cannot set sample format (%s)\n", ::snd_strerror(err));
 		return false;
 	}
-	
+
 	if ((err = ::snd_pcm_hw_params_set_rate(recHandle, hw_params, m_sampleRate, 0)) < 0) {
 		::fprintf(stderr, "Cannot set sample rate (%s)\n", ::snd_strerror(err));
 		return false;
 	}
-	
+
 	unsigned int recChannels = 1U;
-	
+
 	if ((err = ::snd_pcm_hw_params_set_channels(recHandle, hw_params, 1)) < 0) {
 		recChannels = 2U;
 
@@ -330,7 +330,7 @@ bool CSoundCardReaderWriter::open()
 			return false;
 		}
 	}
-	
+
 	if ((err = ::snd_pcm_hw_params(recHandle, hw_params)) < 0) {
 		::fprintf(stderr, "Cannot set parameters (%s)\n", ::snd_strerror(err));
 		return false;
@@ -456,25 +456,25 @@ CSoundCardWriter::~CSoundCardWriter()
 void CSoundCardWriter::entry()
 {
 	while (!m_killed) {
-		unsigned int nSamples = 2U * m_blockSize;
+		int nSamples = 2U * m_blockSize;
 		m_callback->writeCallback(m_buffer, nSamples);
 
 		if (nSamples == 0U) {
 			sleep(5UL);
 		} else {
 			if (m_channels == 1U) {
-				for (unsigned int n = 0U; n < nSamples; n++)
+				for (int n = 0U; n < nSamples; n++)
 					m_samples[n] = short(m_buffer[n] * 32767.0F);
 			} else {
-				unsigned int i = 0U;
-				for (unsigned int n = 0U; n < nSamples; n++) {
+				int i = 0U;
+				for (int n = 0U; n < nSamples; n++) {
 					short sample = short(m_buffer[n] * 32767.0F);
 					m_samples[i++] = sample;
 					m_samples[i++] = sample;			// Same value to both channels
 				}
 			}
-        
-			unsigned int offset = 0U;
+
+			int offset = 0U;
 			snd_pcm_sframes_t ret;
 			while ((ret = ::snd_pcm_writei(m_handle, m_samples + offset, nSamples - offset)) != (nSamples - offset)) {
 				if (ret < 0) {
@@ -483,7 +483,7 @@ void CSoundCardWriter::entry()
 
 					::snd_pcm_recover(m_handle, ret, 1);
 				} else {
-					offset += (unsigned int)ret;
+					offset += ret;
 				}
 			}
 		}
